CS-343 - Limited access to backend menus and controllers based on permissions
