+++ csatar-plugins/csatar/controllers/AgeGroups.php
@@ -6,7 +6,7 @@ use BackendMenu;
 class AgeGroups extends Controller
 {
     public $implement = [        'Backend\Behaviors\ListController',        'Backend\Behaviors\FormController',        'Backend\Behaviors\ReorderController'    ];
-    
+
     public $listConfig = 'config_list.yaml';
     public $formConfig = 'config_form.yaml';
     public $reorderConfig = 'config_reorder.yaml';
@@ -16,4 +16,8 @@ class AgeGroups extends Controller
         parent::__construct();
         BackendMenu::setContext('Csatar.Csatar', 'main-menu-item-organization-system-data', 'side-menu-item-ageGroups');
     }
+
+    public $requiredPermissions = [
+        'csatar.manage.data'
+    ];
 }
