+++ csatar-plugins/forms/lang/en/lang.php
@@ -3,6 +3,9 @@
         'name' => 'Form Builder',
         'description' => 'Front end form builder',
         'author' => 'Csatár team',
+        'permissions' => [
+            'manageData' => 'Manage Forms',
+        ],
         'admin' => [
             'form' => [
                 'title' => 'Title',
